Add sounds
Added sounds:
Enemy type1,  player
