+++ Money_Maker/Assets/Enemy.mixer
@@ -1,0 +1,141 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!243 &-3043160176170230222
+AudioMixerGroupController:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_Name: Effects
+  m_AudioMixer: {fileID: 24100000}
+  m_GroupID: 44982ca7fb1b72d46921bdf6b826d42b
+  m_Children: []
+  m_Volume: b20a914dc46586648a389827dca93f6d
+  m_Pitch: 3347bdedb373f5c47ac353f17534f9d6
+  m_Send: 00000000000000000000000000000000
+  m_Effects:
+  - {fileID: 1282174637084319456}
+  m_UserColorIndex: 0
+  m_Mute: 0
+  m_Solo: 0
+  m_BypassEffects: 0
+--- !u!241 &24100000
+AudioMixerController:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_Name: Enemy
+  m_OutputGroup: {fileID: 0}
+  m_MasterGroup: {fileID: 24300002}
+  m_Snapshots:
+  - {fileID: 24500006}
+  m_StartSnapshot: {fileID: 24500006}
+  m_SuspendThreshold: -80
+  m_EnableSuspend: 1
+  m_UpdateMode: 0
+  m_ExposedParameters:
+  - guid: 9d83caee5075c2b4dba3605d0ea60684
+    name: enemyVol
+  m_AudioMixerGroupViews:
+  - guids:
+    - ffe0856b55d3f114bb02a009d463984a
+    - 244b1493c1e2e8649a71ac8d7e2501fe
+    - 44982ca7fb1b72d46921bdf6b826d42b
+    name: View
+  m_CurrentViewIndex: 0
+  m_TargetSnapshot: {fileID: 24500006}
+--- !u!243 &24300002
+AudioMixerGroupController:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_Name: Master
+  m_AudioMixer: {fileID: 24100000}
+  m_GroupID: ffe0856b55d3f114bb02a009d463984a
+  m_Children:
+  - {fileID: 1987742270892884046}
+  - {fileID: -3043160176170230222}
+  m_Volume: 9d83caee5075c2b4dba3605d0ea60684
+  m_Pitch: e0207aaca31cee14d8e2b6375fa3adcb
+  m_Send: 00000000000000000000000000000000
+  m_Effects:
+  - {fileID: 24400004}
+  m_UserColorIndex: 0
+  m_Mute: 0
+  m_Solo: 0
+  m_BypassEffects: 0
+--- !u!244 &24400004
+AudioMixerEffectController:
+  m_ObjectHideFlags: 3
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_Name: 
+  m_EffectID: 494dc9171d4e106449b533639c26ec2f
+  m_EffectName: Attenuation
+  m_MixLevel: a490fa30477dbaf4b8e48fe58872803c
+  m_Parameters: []
+  m_SendTarget: {fileID: 0}
+  m_EnableWetMix: 0
+  m_Bypass: 0
+--- !u!245 &24500006
+AudioMixerSnapshotController:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_Name: Snapshot
+  m_AudioMixer: {fileID: 24100000}
+  m_SnapshotID: 947ddbef0cb5ff1448299f4b9191db53
+  m_FloatValues:
+    9a98b5c333a0faa45b12d44796316f57: 0
+  m_TransitionOverrides: {}
+--- !u!244 &1282174637084319456
+AudioMixerEffectController:
+  m_ObjectHideFlags: 3
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_Name: 
+  m_EffectID: bc8fae581478ec449b0e1b91270d1186
+  m_EffectName: Attenuation
+  m_MixLevel: d5ab46838303b9146a49e6dc9802a06d
+  m_Parameters: []
+  m_SendTarget: {fileID: 0}
+  m_EnableWetMix: 0
+  m_Bypass: 0
+--- !u!243 &1987742270892884046
+AudioMixerGroupController:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_Name: Enemy
+  m_AudioMixer: {fileID: 24100000}
+  m_GroupID: 244b1493c1e2e8649a71ac8d7e2501fe
+  m_Children: []
+  m_Volume: 9a98b5c333a0faa45b12d44796316f57
+  m_Pitch: bd8f2bbe35ef24f4cacf925e4a3cfa2e
+  m_Send: 00000000000000000000000000000000
+  m_Effects:
+  - {fileID: 3829388588051154282}
+  m_UserColorIndex: 0
+  m_Mute: 0
+  m_Solo: 0
+  m_BypassEffects: 0
+--- !u!244 &3829388588051154282
+AudioMixerEffectController:
+  m_ObjectHideFlags: 3
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_Name: 
+  m_EffectID: 6fbfcc9659bf2b340a2cdb88a5222f69
+  m_EffectName: Attenuation
+  m_MixLevel: c0fc55c22ae1d254e9ccce9f09eec213
+  m_Parameters: []
+  m_SendTarget: {fileID: 0}
+  m_EnableWetMix: 0
+  m_Bypass: 0
