+++ Money_Maker/Assets/Enemy.mixer.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 88f9154ac5404444eb9e26cb0890d10c
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 24100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
