+++ Money_Maker/Assets/Player.mixer.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 218ef60396cc8134581708db39e40cf1
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 24100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
