+++ Money_Maker/Assets/Scripts/Player/Shoot.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 71d470a03b0ac1443bc046ad1df0703a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
