+++ Money_Maker/Assets/Scripts/Player/SoundsHandler.cs
@@ -1,0 +1,17 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class SoundsHandler : MonoBehaviour
+{
+
+    public void PlaySound(AudioSource audio)
+    {
+        audio.Play();
+    }
+
+    public void StopSound(AudioSource audio)
+    {
+        audio.Stop();
+    }
+}
