+++ Money_Maker/Assets/Scripts/Player/SoundsHandler.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e50dbc4e166d06f4f97e0630e778a5b1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
