--- Money_Maker/Assets/Scripts/Players.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 41dea2087f6cfab49a8ba04e75f0c79e
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
