--- Money_Maker/Assets/Scripts/Players/LookAtMouseAndRotate.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 94e2aa28b68eb994d8b90085efd3acd5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
