--- Money_Maker/Assets/Scripts/Players/PlayerControl.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 29fc6375192af914f8a70bd799a2e3b8
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
