+++ Money_Maker/Assets/Scripts/Sound.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f2d506ad6e7691445af86fbdff2cbfa4
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
