+++ Money_Maker/Assets/Scripts/Sound/SoundManager.cs
@@ -1,0 +1,7 @@
+using UnityEngine;
+using UnityEngine.Audio;
+
+public class SoundManager : MonoBehaviour
+{
+    
+}
