+++ Money_Maker/Assets/Sound/Enemy.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d4a84e984ce2b3b45a4d3df7e76fa983
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
