+++ Money_Maker/Assets/Sound/Enemy/EnemyTypeOneSteps.mp3.meta
@@ -1,0 +1,22 @@
+fileFormatVersion: 2
+guid: 650b48850f3a7744dba8850918a7e1c8
+AudioImporter:
+  externalObjects: {}
+  serializedVersion: 6
+  defaultSettings:
+    loadType: 0
+    sampleRateSetting: 0
+    sampleRateOverride: 44100
+    compressionFormat: 1
+    quality: 1
+    conversionMode: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  normalize: 1
+  preloadAudioData: 1
+  loadInBackground: 0
+  ambisonic: 0
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
