+++ Money_Maker/Assets/Sound/Player.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 61682ff66700a3e4b8bd9359928249dc
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
