+++ Money_Maker/Assets/Sound/UI.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 56dd27fc807e1df448e3fe0b30462918
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
