--- Money_Maker/Logs/shadercompiler-UnityShaderCompiler.exe1.log
@@ -1,3 +1,0 @@
-Base path: 'C:/Program Files/Unity/Hub/Editor/2020.3.4f1/Editor/Data', plugins path 'C:/Program Files/Unity/Hub/Editor/2020.3.4f1/Editor/Data/PlaybackEngines'
-Cmd: initializeCompiler
-
